reorganize functions according to domain
