+++ cli/internal/app/onestop/onestop.go
@@ -1,23 +1,14 @@
 package onestop
 
 import (
-	"gopkg.in/h2non/gentleman.v2"
-	"github.com/spf13/viper"
-	"strings"
 	"github.com/danielgtaylor/openapi-cli-generator/cli"
 	"github.com/cedardevs/onestop/cli/internal/pkg/flags"
-	"github.com/cedardevs/onestop/cli/internal/pkg/utils"
+	"github.com/cedardevs/onestop/cli/internal/pkg/middleware/onestop"
 )
 
 const SearchCollectionCmd = "searchcollection"
 const SearchGranuleCmd = "searchgranule"
 const SearchFlattenedGranuleCmd = "searchflattenedgranule"
-
-func InjectMiddleware(){
-	cli.RegisterBefore(SearchCollectionCmd, ParseOneStopRequestFlags)
-	cli.RegisterBefore(SearchGranuleCmd, ParseOneStopRequestFlags)
-	cli.RegisterBefore(SearchFlattenedGranuleCmd, ParseOneStopRequestFlags)
-}
 
 //flags in flags.go
 func SetOneStopFlags() {
@@ -51,22 +42,8 @@ func SetOneStopFlags() {
 	cli.AddFlag(SearchFlattenedGranuleCmd, flags.StartTimeFlag, flags.StartTimeShortFlag, flags.StartTimeScdrDescription, "")
 }
 
-//this function is pre-request RegisterBefore
-func ParseOneStopRequestFlags(cmd string, params *viper.Viper, req *gentleman.Request) {
-	filters := []string{}
-	queries := []string{}
-
-	dateTimeFilter := utils.ParseDate(params)
-	filters = append(filters, dateTimeFilter...)
-	startEndTimeFilter := utils.ParseStartAndEndTime(params)
-	filters = append(filters, startEndTimeFilter...)
-	geoSpatialFilter := utils.ParsePolygon(params)
-	filters = append(filters, geoSpatialFilter...)
-	query := utils.ParseTextQuery(params)
-	queries = append(queries, query...)
-	requestMeta := utils.ParseRequestMeta(params)
-	if len(queries) > 0 || len(filters) > 0 {
-		req.AddHeader("content-type", "application/json")
-		req.BodyString("{\"filters\":[" + strings.Join(filters, ", ") + "], \"queries\":[" + strings.Join(queries, ", ") + "]," + requestMeta + "}")
-	}
+func InjectMiddleware(){
+	cli.RegisterBefore(SearchCollectionCmd, middleware.ParseOneStopRequestFlags)
+	cli.RegisterBefore(SearchGranuleCmd, middleware.ParseOneStopRequestFlags)
+	cli.RegisterBefore(SearchFlattenedGranuleCmd, middleware.ParseOneStopRequestFlags)
 }
