+++ cli/internal/app/scdr/scdr.go
@@ -11,8 +11,7 @@ import (
 	"strings"
 	"github.com/cedardevs/onestop/cli/internal/pkg/flags"
 	"github.com/cedardevs/onestop/cli/internal/app/generated"
-	"github.com/cedardevs/onestop/cli/internal/pkg/utils"
-	"github.com/cedardevs/onestop/cli/internal/pkg/middleware"
+	"github.com/cedardevs/onestop/cli/internal/pkg/middleware/scdr"
 )
 
 const ScdrFileCmd = "scdr-files"
@@ -23,6 +22,43 @@ scdr-files --area "POLYGON(( 22.686768 34.051522, 30.606537 34.051522, 30.606537
 scdr-files --date 10/01
 scdr-files --stime "March 31st 2003 at 17:30" --etime "2003-04-01 10:32:49"
 `
+
+func SetScdrFlags() {
+	//flags are in flags.go
+	cli.AddFlag(ScdrFileCmd, flags.DateFilterFlag, flags.DateFilterShortFlag, flags.DateDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.TypeFlag, flags.TypeShortFlag, flags.TypeDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.SpatialFilterFlag, flags.SpatialFilterShortFlag, flags.AreaDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.StartTimeFlag, flags.StartTimeShortFlag, flags.StartTimeDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.StartTimeScdrFlag, "", flags.StartTimeScdrDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.EndTimeFlag, flags.EndTimeShortFlag, flags.EndTimeDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.EndTimeScdrFlag, "", flags.EndTimeScdrDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.AvailableFlag, flags.AvailableShortFlag, flags.AvailableDescription, false)
+	cli.AddFlag(ScdrFileCmd, flags.MetadataFlag, flags.MetadataShortFlag, flags.MetadataDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.FileFlag, flags.FileShortFlag, flags.FileFlagDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.ReFileFlag, flags.ReFileShortFlag, flags.RegexDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.SatnameFlag, "", flags.SatnameDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.YearFlag, flags.YearShortFlag, flags.YearDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.KeywordFlag, flags.KeywordShortFlag, flags.KeywordDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.GapFlag, flags.GapShortFlag, flags.GapDescription, "")
+
+//not scdr-files specific
+	cli.AddFlag(ScdrFileCmd, flags.MaxFlag, flags.MaxShortFlag, flags.MaxDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.OffsetFlag, flags.OffsetShortFlag, flags.OffsetDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.TextQueryFlag, flags.TextQueryShortFlag, flags.QueryDescription, "")
+	cli.AddFlag(ScdrFileCmd, flags.CloudServerFlag, flags.CloudServerShortFlag, flags.CloudServerDescription, false)
+	cli.AddFlag(ScdrFileCmd, flags.TestServerFlag, flags.TestServerShortFlag, flags.TestServerDescription, false)
+	cli.AddFlag(ScdrFileCmd, flags.SortFlag, flags.SortShortFlag, flags.SortDescription, "")
+
+}
+
+func InjectMiddleware(){
+	//ParseScdrRequestFlags in parsing-util.go
+	cli.RegisterBefore(ScdrFileCmd, middleware.ParseScdrRequestFlags)
+	cli.RegisterAfter(ScdrFileCmd, func(cmd string, params *viper.Viper, resp *gentleman.Response, data interface{}) interface{} {
+		scdrResp := middleware.MarshalScdrResponse(params, data)
+		return scdrResp
+	})
+}
 
 func ScdrRegister() {
 	root := cli.Root
@@ -80,7 +116,7 @@ func ScdrRegister() {
 func ScdrSearch(params *viper.Viper, body string) (*gentleman.Response, map[string]interface{}, error) {
 	handlerPath := "scdr-files"
 
-	params = translateArgs(params)
+	params = middleware.TranslateArgs(params)
 	req := buildRequest(params)
 
 	if body != "" {
@@ -112,18 +148,6 @@ func ScdrSearch(params *viper.Viper, body string) (*gentleman.Response, map[stri
 	return resp, decoded, nil
 }
 
-func translateArgs(params *viper.Viper) *viper.Viper {
-	typeArg := params.GetString(flags.TypeFlag)
-	err := viper.ReadInConfig()
-	if err != nil {
-		return params
-	}
-	scdrTypeIds := viper.Get("scdr-types").(map[string]interface{})
-	uuid := scdrTypeIds[strings.ToLower(typeArg)]
-	params.Set("type", uuid)
-	return params
-}
-
 func buildRequest(params *viper.Viper) *gentleman.Request {
 	server := viper.GetString("server")
 	if params.GetString("test") == "true" {
@@ -143,7 +167,7 @@ func buildRequest(params *viper.Viper) *gentleman.Request {
 
   /// default endpoint: search/flattened-granule
 	// /collection/{id} if we have type to get count
-	endpoint := determineEndpoint(params, isSummaryWithType)
+	endpoint := middleware.DetermineEndpoint(params, isSummaryWithType)
 
 	url := server + endpoint
 
@@ -158,18 +182,6 @@ func buildRequest(params *viper.Viper) *gentleman.Request {
 	return req
 }
 
-func determineEndpoint(params *viper.Viper, isSummaryWithType bool) string {
-	endpoint := "/search/flattened-granule"
-
-	if isSummaryWithType {
-		collectionId := params.GetString(flags.TypeFlag)
-		endpoint = "/collection/" + collectionId
-	} else if params.GetString(flags.AvailableFlag) == "true" {
-		endpoint = "/search/collection"
-	}
-	return endpoint
-}
-
 func scdrOutputFormatAndPrint(params *viper.Viper, decoded map[string]interface{}) {
 	if output, ok := decoded["scdr-output"].([]string); ok {
 		for _, row := range output {
@@ -177,83 +189,5 @@ func scdrOutputFormatAndPrint(params *viper.Viper, decoded map[string]interface{
 		}
 	} else {
 		fmt.Println("No results")
-	}
-}
-
-func InjectMiddleware(){
-	//ParseScdrRequestFlags in parsing-util.go
-	cli.RegisterBefore(ScdrFileCmd, ParseScdrRequestFlags)
-	cli.RegisterAfter(ScdrFileCmd, func(cmd string, params *viper.Viper, resp *gentleman.Response, data interface{}) interface{} {
-		scdrResp := middleware.MarshalScdrResponse(params, data)
-		return scdrResp
-	})
-}
-
-func SetScdrFlags() {
-	//flags are in flags.go
-	cli.AddFlag(ScdrFileCmd, flags.DateFilterFlag, flags.DateFilterShortFlag, flags.DateDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.TypeFlag, flags.TypeShortFlag, flags.TypeDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.SpatialFilterFlag, flags.SpatialFilterShortFlag, flags.AreaDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.StartTimeFlag, flags.StartTimeShortFlag, flags.StartTimeDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.StartTimeScdrFlag, "", flags.StartTimeScdrDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.EndTimeFlag, flags.EndTimeShortFlag, flags.EndTimeDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.EndTimeScdrFlag, "", flags.EndTimeScdrDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.AvailableFlag, flags.AvailableShortFlag, flags.AvailableDescription, false)
-	cli.AddFlag(ScdrFileCmd, flags.MetadataFlag, flags.MetadataShortFlag, flags.MetadataDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.FileFlag, flags.FileShortFlag, flags.FileFlagDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.ReFileFlag, flags.ReFileShortFlag, flags.RegexDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.SatnameFlag, "", flags.SatnameDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.YearFlag, flags.YearShortFlag, flags.YearDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.KeywordFlag, flags.KeywordShortFlag, flags.KeywordDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.GapFlag, flags.GapShortFlag, flags.GapDescription, "")
-
-//not scdr-files specific
-	cli.AddFlag(ScdrFileCmd, flags.MaxFlag, flags.MaxShortFlag, flags.MaxDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.OffsetFlag, flags.OffsetShortFlag, flags.OffsetDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.TextQueryFlag, flags.TextQueryShortFlag, flags.QueryDescription, "")
-	cli.AddFlag(ScdrFileCmd, flags.CloudServerFlag, flags.CloudServerShortFlag, flags.CloudServerDescription, false)
-	cli.AddFlag(ScdrFileCmd, flags.TestServerFlag, flags.TestServerShortFlag, flags.TestServerDescription, false)
-	cli.AddFlag(ScdrFileCmd, flags.SortFlag, flags.SortShortFlag, flags.SortDescription, "")
-
-}
-
-func ParseScdrRequestFlags(cmd string, params *viper.Viper, req *gentleman.Request) {
-
-	//apply a default filter for STAR
-	filters := []string{"{\"type\":\"facet\",\"name\":\"dataCenters\",\"values\":[\"DOC/NOAA/NESDIS/STAR > Center for Satellite Applications and Research, NESDIS, NOAA, U.S. Department of Commerce\"]}"}
-	queries := []string{}
-
-	// isSummaryWithType := params.GetString(AvailableFlag) == "true" && len(params.GetString("type")) > 0
-
-	collectionIdFilter := utils.ParseTypeFlag(params)
-	filters = append(filters, collectionIdFilter...)
-	// datacenterFilter := parseAvailableFlag(params)
-	// filters = append(filters, datacenterFilter...)
-	dateTimeFilter := utils.ParseDate(params)
-	filters = append(filters, dateTimeFilter...)
-	yearFilter := utils.ParseYear(params)
-	filters = append(filters, yearFilter...)
-	startEndTimeFilter := utils.ParseStartAndEndTime(params)
-	filters = append(filters, startEndTimeFilter...)
-	geoSpatialFilter := utils.ParsePolygon(params)
-	filters = append(filters, geoSpatialFilter...)
-
-	satnameQuery := utils.ParseSatName(params)
-	queries = append(queries, satnameQuery...)
-	fileNameQuery := utils.ParseFileName(params)
-	queries = append(queries, fileNameQuery...)
-	refileNameQuery := utils.ParseRegexFileName(params)
-	queries = append(queries, refileNameQuery...)
-	query := utils.ParseTextQuery(params)
-	queries = append(queries, query...)
-	keyWordFilter := utils.ParseKeyword(params)
-	queries = append(queries, keyWordFilter...)
-	requestMeta := utils.ParseRequestMeta(params)
-
-	parseSort := utils.ParseSort(params)
-
-	if len(queries) > 0 || len(filters) > 0 {
-		req.AddHeader("content-type", "application/json")
-		req.BodyString("{\"summary\":false, \"sort\":[" + parseSort + "], \"filters\":[" + strings.Join(filters, ", ") + "], \"queries\":[" + strings.Join(queries, ", ") + "]," + requestMeta + "}")
 	}
 }
